Move evaluator settings (type and global) to separate file/header
